+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+### Server
+- Feat: Mastodon 等からの FEP-044f 引用リクエスト (QuoteRequest) を受け付け、無条件に引用を許可して QuoteAuthorization を応答するように (#776)
+
 ## 2026.6.0
 
 ### General